Split implementation, external sink in driver.go, translation to client in client.go and structs in types.go
